Playlist track position reflects chain track position

+++ BayWave/src/main/java/com/BayWave/Tables/ChainTrackTable.java
@@ -26,12 +26,20 @@ public class ChainTrackTable {
             // check if track is already in a chain in that playlist
             // get playlist id
             int plyId = TableUtil.getPlaylistIdOfChain(connection, chainId);
+            if (plyId == -1) {
+                System.err.println("Playlist ID not found");
+                return;
+            }
             int trackId = TableUtil.getTrackID(connection, artist, album, track);
+            if (trackId == -1) {
+                System.err.println("Track ID not found");
+                return;
+            }
             PreparedStatement ps = connection.prepareStatement("SELECT * FROM CHAIN_ WHERE ply_id=?");
             ps.setInt(1, plyId);
             ResultSet rs = ps.executeQuery();
             if (!rs.isBeforeFirst()) {
-                System.out.println("No chain found associated with playlist");
+                System.err.println("No chain found associated with playlist");
                 return;
             }
             while (rs.next()) {
@@ -39,13 +47,13 @@ public class ChainTrackTable {
                 ps.setInt(1, rs.getInt("chn_id"));
                 ResultSet rs2 = ps.executeQuery();
                 if (!rs2.isBeforeFirst()) {
-                    System.out.println("Chain not found");
+                    System.err.println("Chain not found");
                     return;
                 }
                 while (rs2.next()) {
                     int currTrackId = rs2.getInt("trk_id");
                     if (currTrackId == trackId) {
-                        System.out.println("Track already in chain within playlist");
+                        System.err.println("Track already in chain within playlist");
                         return;
                     }
                 }
@@ -56,9 +64,85 @@ public class ChainTrackTable {
             ps.setInt(2, trackId);
             int result = ps.executeUpdate();
             if (result == 0) {
-                System.out.println("CHAIN_TRACK not added");
+                System.err.println("CHAIN_TRACK not added");
                 return;
             }
+
+            // update track's position in playlist
+            // first, obtain highest playlist position within the chain, excluding the one being added
+            ps = connection.prepareStatement("SELECT * FROM CHAIN_TRACK WHERE chn_id=?");
+            ps.setInt(1, chainId);
+            rs = ps.executeQuery();
+            if (!rs.isBeforeFirst()) {
+                System.err.println("Chain tracks not found");
+                return;
+            }
+            int highestPos = -1;
+            int currPos = -1;
+            while (rs.next()) {
+                int currTrackId = rs.getInt("trk_id");
+                ps = connection.prepareStatement("SELECT * FROM PLAYLIST_TRACK WHERE ply_id=? AND trk_id=?");
+                ps.setInt(1, plyId);
+                ps.setInt(2, currTrackId);
+                ResultSet rs2 = ps.executeQuery();
+                if (!rs2.isBeforeFirst()) {
+                    System.err.println("Playlist track not found");
+                    return;
+                }
+                rs2.next();
+                if (currTrackId != trackId) {
+                    int currPlaylistPos = rs2.getInt("ply_trk_pos");
+                    if (currPlaylistPos > highestPos) {
+                        highestPos = currPlaylistPos;
+                    }
+                }
+                else {
+                    currPos = rs2.getInt("ply_trk_pos");
+                }
+            }
+            if (highestPos == -1) {
+                System.err.println("Highest track position not found");
+                return;
+            }
+            if (currPos == -1) {
+                System.err.println("Current track position not found");
+                return;
+            }
+            // get user and playlist names
+            String playlist = TableUtil.getPlaylistNameFromId(connection, plyId);
+            if (playlist == null) {
+                System.err.println("Playlist name not found");
+                return;
+            }
+            ps = connection.prepareStatement("SELECT usr_id FROM PLAYLIST WHERE ply_id=?");
+            ps.setInt(1, plyId);
+            rs = ps.executeQuery();
+            if (!rs.isBeforeFirst()) {
+                System.err.println("User ID not found");
+                return;
+            }
+            rs.next();
+            String user = TableUtil.getUsernameFromId(connection, rs.getInt("usr_id"));
+            if (user == null) {
+                System.err.println("Username not found");
+                return;
+            }
+
+            if (highestPos < currPos) { // insert track 2 at plyPos1 + 1
+                PlaylistTrackTable.insertAtPosition(connection, user, playlist, artist, album, track, highestPos + 1);
+            }
+            else { // insert track 2 at plyPos1
+                PlaylistTrackTable.insertAtPosition(connection, user, playlist, artist, album, track, highestPos);
+            }
+
+            /*
+            When a track is added to a chain, it should follow the same rules as the second
+            track when its position was being changed in the playlist. If the track comes
+            after the chain, it should be inserted at the position of the last track
+            in the chain + 1. But if the track comes before the chain, it should be
+            inserted at the position of the last track in the chain.
+             */
+
             connection.commit();
             System.out.println("CHAIN_TRACK added");
         }
@@ -70,6 +154,7 @@ public class ChainTrackTable {
     /**
      * Removes a specified track from a specified chain provided that the chain has at least
      * three tracks, enforcing that the chain will always have at least two tracks.
+     * Track's position in playlist will be shifted after the chain.
      */
     public static void removeTrack(Connection connection, int chainId, String artist, String album, String track) throws SQLException {
         try {
@@ -83,22 +168,89 @@ public class ChainTrackTable {
                 trackCount++;
             }
             if (trackCount < 3) {
-                System.out.println("Chain must always have at least two tracks, not deleted");
+                System.err.println("Chain must always have at least two tracks, not deleted");
                 return;
             }
             int trackId = TableUtil.getTrackID(connection, artist, album, track);
             if (trackId == -1) {
-                System.out.println("Track not found");
+                System.err.println("Track not found");
                 return;
             }
+            int plyId = TableUtil.getPlaylistIdOfChain(connection, chainId);
+            if (plyId == -1) {
+                System.err.println("Playlist ID not found");
+                return;
+            }
+
             ps = connection.prepareStatement("DELETE FROM CHAIN_TRACK WHERE chn_id=? AND trk_id=?");
             ps.setInt(1, chainId);
             ps.setInt(2, trackId);
             int result = ps.executeUpdate();
             if (result == 0) {
-                System.out.println("Track not removed");
+                System.err.println("Track not removed");
                 return;
             }
+            // update track's position in playlist
+
+            ps = connection.prepareStatement("SELECT * FROM CHAIN_TRACK WHERE chn_id=?");
+            ps.setInt(1, chainId);
+            rs = ps.executeQuery();
+            if (!rs.isBeforeFirst()) {
+                System.err.println("Chain tracks not found");
+                return;
+            }
+            int highestPos = -1;
+            while (rs.next()) {
+                int currTrackId = rs.getInt("trk_id");
+                ps = connection.prepareStatement("SELECT * FROM PLAYLIST_TRACK WHERE ply_id=? AND trk_id=?");
+                ps.setInt(1, plyId);
+                ps.setInt(2, currTrackId);
+                ResultSet rs2 = ps.executeQuery();
+                if (!rs2.isBeforeFirst()) {
+                    System.err.println("Playlist track not found");
+                    return;
+                }
+                rs2.next();
+                if (currTrackId != trackId) {
+                    int currPlaylistPos = rs2.getInt("ply_trk_pos");
+                    if (currPlaylistPos > highestPos) {
+                        highestPos = currPlaylistPos;
+                    }
+                }
+            }
+            if (highestPos == -1) {
+                System.err.println("Highest track position not found");
+                return;
+            }
+
+            // get user and playlist names
+            String playlist = TableUtil.getPlaylistNameFromId(connection, plyId);
+            if (playlist == null) {
+                System.err.println("Playlist name not found");
+                return;
+            }
+            ps = connection.prepareStatement("SELECT usr_id FROM PLAYLIST WHERE ply_id=?");
+            ps.setInt(1, plyId);
+            rs = ps.executeQuery();
+            if (!rs.isBeforeFirst()) {
+                System.err.println("User ID not found");
+                return;
+            }
+            rs.next();
+            String user = TableUtil.getUsernameFromId(connection, rs.getInt("usr_id"));
+            if (user == null) {
+                System.err.println("Username not found");
+                return;
+            }
+
+            PlaylistTrackTable.insertAtPosition(connection, user, playlist, artist, album, track, highestPos);
+
+            /*
+            If a track is removed from a chain, it should be re-positioned to come after
+            the chain. The current position of the track is guaranteed to come before
+            that position, so the track should be inserted at the position of the last
+            track in the chain.
+             */
             connection.commit();
             System.out.println("Track removed");
         }
@@ -107,13 +259,15 @@ public class ChainTrackTable {
         }
     }
 
+    // any of these position changes should be mirrored in the playlist
+
     public static void swapPosition(Connection connection, int chainId, String artist, String album, String track, int newPos) throws SQLException {
         try {
             Reset.lock.lock();
             // get track id
             int trkId = TableUtil.getTrackID(connection, artist, album, track);
             if (trkId == -1) {
-                System.out.println("Track not found");
+                System.err.println("Track not found");
                 return;
             }
             // check that a track exists with the new position
@@ -123,7 +277,7 @@ public class ChainTrackTable {
             ps.setInt(2, trkId);
             ResultSet rs = ps.executeQuery();
             if (!rs.isBeforeFirst()) {
-                System.out.println("Track not found");
+                System.err.println("Track not found");
                 return;
             }
             rs.next();
@@ -135,14 +289,14 @@ public class ChainTrackTable {
             ps.setInt(2, newPos);
             rs = ps.executeQuery();
             if (!rs.isBeforeFirst()) {
-                System.out.println("Track position not found");
+                System.err.println("Track position not found");
                 return;
             }
             rs.next();
             int trkPos2 = rs.getInt("chn_trk_pos");
             int trkId2 = rs.getInt("trk_id");
             if (trkPos1 == trkPos2) {
-                System.out.println("Track already at position specified");
+                System.err.println("Track already at position specified");
                 return;
             }
             ps = connection.prepareStatement("UPDATE CHAIN_TRACK SET chn_trk_pos=? WHERE trk_id=?");
@@ -150,7 +304,7 @@ public class ChainTrackTable {
             ps.setInt(2, trkId1);
             int updated = ps.executeUpdate();
             if (updated == 0) {
-                System.out.println("Track position not updated");
+                System.err.println("Track position not updated");
                 return;
             }
             ps = connection.prepareStatement("UPDATE CHAIN_TRACK SET chn_trk_pos=? WHERE trk_id=?");
@@ -158,7 +312,7 @@ public class ChainTrackTable {
             ps.setInt(2, trkId2);
             int updated2 = ps.executeUpdate();
             if (updated2 == 0) {
-                System.out.println("Track position not updated");
+                System.err.println("Track position not updated");
                 return;
             }
             connection.commit();
@@ -172,7 +326,7 @@ public class ChainTrackTable {
     public static void insertAtPosition(Connection connection, int chainId, String artist, String album, String track, int newPos) throws SQLException {
         int trkId = TableUtil.getTrackID(connection, artist, album, track);
         if (trkId == -1) {
-            System.out.println("Track not found");
+            System.err.println("Track not found");
             return;
         }
         PreparedStatement ps = connection.prepareStatement("SELECT * FROM CHAIN_TRACK WHERE chn_id=? and trk_id=?");
@@ -180,13 +334,13 @@ public class ChainTrackTable {
         ps.setInt(2, trkId);
         ResultSet rs = ps.executeQuery();
         if (!rs.isBeforeFirst()) {
-            System.out.println("Track position not found");
+            System.err.println("Track position not found");
             return;
         }
         rs.next();
         int currPos = rs.getInt("chn_trk_pos");
         if (currPos == newPos) {
-            System.out.println("Track already at position specified");
+            System.err.println("Track already at position specified");
             return;
         }
         int delta;
@@ -202,7 +356,7 @@ public class ChainTrackTable {
         ps.setInt(2, newPos);
         rs = ps.executeQuery();
         if (!rs.isBeforeFirst()) {
-            System.out.println("Track position not found");
+            System.err.println("Track position not found");
             return;
         }
         rs.next();
@@ -212,25 +366,22 @@ public class ChainTrackTable {
         ps = connection.prepareStatement("SELECT * FROM CHAIN_TRACK WHERE chn_id=? AND chn_trk_pos>=? AND chn_trk_pos<=?");
         ps.setInt(1, chainId);
         if (delta == -1) {
-            System.out.println("DELTA -1");
             ps.setInt(2, currPos);
             ps.setInt(3, newPos);
         }
         else {
-            System.out.println("DELTA 1");
             ps.setInt(2, newPos);
             ps.setInt(3, currPos);
         }
         rs = ps.executeQuery();
         if (!rs.isBeforeFirst()) {
-            System.out.println("Error obtaining track positions");
+            System.err.println("Error obtaining track positions");
             return;
         }
         while (rs.next()) {
             int thisId = rs.getInt("trk_id");
             int thisPos = rs.getInt("chn_trk_pos");
             if (thisId != trkId) {
-                System.out.println("UPDATING TRACK");
                 ps = connection.prepareStatement("UPDATE CHAIN_TRACK SET chn_trk_pos=? WHERE chn_id=? AND trk_id=?");
                 int updatedPos = thisPos + delta;
                 ps.setInt(1, updatedPos);
@@ -238,7 +389,7 @@ public class ChainTrackTable {
                 ps.setInt(3, thisId);
                 int result = ps.executeUpdate();
                 if (result == 0) {
-                    System.out.println("Track position update failed");
+                    System.err.println("Track position update failed");
                     return;
                 }
             }
@@ -250,7 +401,7 @@ public class ChainTrackTable {
         ps.setInt(3, trkId);
         int result = ps.executeUpdate();
         if (result == 0) {
-            System.out.println("Track position not updated");
+            System.err.println("Track position not updated");
             return;
         }
         System.out.println("Track positions updated");
@@ -263,7 +414,7 @@ public class ChainTrackTable {
         ps.setInt(1, chainId);
         ResultSet rs = ps.executeQuery();
         if (rs.isBeforeFirst()) {
-            System.out.println("Chain not found");
+            System.err.println("Chain not found");
             return null;
         }
         return TableUtil.getTable(rs);
